refactor: Rename `…compile/events/lib/ensureApiGatewayCloudWatchRole.js`
