+++ lib/plugins/aws/package/compile/events/apiGateway/lib/stage.js
@@ -5,7 +5,7 @@
 // NOTE: --> Keep this file in sync with ./hack/updateStage.js
 
 const _ = require('lodash');
-const ensureApiGatewayCloudWatchRole = require('../../lib/ensureApiGatewayCloudWatchRole');
+const ensureApiGatewayCloudWatchRole = require('../../lib/ensure-api-gateway-cloud-watch-role');
 
 module.exports = {
   async compileStage() {
