+++ lib/plugins/aws/package/compile/events/lib/ensure-api-gateway-cloud-watch-role.js
