+++ lib/plugins/aws/package/compile/events/websockets/lib/stage.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const ensureApiGatewayCloudWatchRole = require('../../lib/ensureApiGatewayCloudWatchRole');
+const ensureApiGatewayCloudWatchRole = require('../../lib/ensure-api-gateway-cloud-watch-role');
 const ServerlessError = require('../../../../../../../serverless-error');
 
 const defaultLogLevel = 'INFO';
